NO SUBSTANTIVE CHANGES!   All the lines changed are just some unnecessary reordering of the affixes as a side effect of the change I'm about to describe:
Changed code for stripping redundancies in affixes -- used
to just look for one word having a "subset" of the other
and wiped it.  There were three words not handled by this:

fuaim/BMDUS vs. fuaim/USA
obainn/HBMDU vs. obainn/E
sáinn/ST vs. sáinn/E

(and sin/S not wiping "sin" because they were too far apart in
the list -- fixed that too).

Changed the code so that everything goes on one line.
One consequence was that I had to change the BDMT flags to
NOT combine with (entirely verbal) affixes!

Result was supposed to be no change to aspell.txt.   This is
basically true, but I uncovered the fact that the "H" flag
is not always added to, e.g., verbs flagged with "E".
(since hobadar, etc. appeared when the lines above were melded)

Will fix this later.