Changed code so that "H" flag is only used for imperative forms of verbs.  When the verb starts with a vowel, this are all written individually to gaeilge.raw with /H, otherwise (consonants) the affix file takes care of the imperatives (saving space).
Had to modify index-en.html slightly since I'd claimed "over 100"
forms are added for a verb starting with "i" -- this is more like
80 after the incorrect "h" forms are taken out.

The 3rd person imperatives and autonomous imperatives are used
several times in the Bible, so I included them.  Can't find any
instance of a 1st person imperative, so I'm not including those.